+++ src/baseConfig.js
@@ -2,6 +2,7 @@ import babelParser from '@babel/eslint-parser';
 import babelPlugin from '@babel/eslint-plugin';
 import envPreset from '@babel/preset-env';
 import importAttributesSyntaxPlugin from '@babel/plugin-syntax-import-attributes';
+import stylisticPlugin from '@stylistic/eslint-plugin';
 import typescriptParser from '@typescript-eslint/parser';
 import typescript from '@typescript-eslint/eslint-plugin';
 import globals from 'globals';
@@ -40,7 +41,8 @@ export default [
 		],
 
 		plugins: {
-			'@babel': babelPlugin
+			'@babel': babelPlugin,
+			'@stylistic': stylisticPlugin
 		},
 
 		languageOptions: {